models/v2/pyproject.toml
[project]
name = "scada_tcn"
version = "0.0.0"
requires-python = ">=3.10"

dependencies = [
  "numpy>=2.4.1",
  "pandas>=3.0.0",
  "pyarrow>=23.0.0",
  "torch>=2.0",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "tqdm>=4.67.1",
  "scikit-learn>=1.8.0",
]
